requirements.txt
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.5
django-dotenv==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.7.1
hyperlink==21.0.0
identify==2.5.1
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.0
nodeenv==1.6.0
openpyxl==3.0.10
parsel==1.6.0
platformdirs==2.5.2
pre-commit==2.19.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pytz==2022.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.0
toml==0.10.2
Twisted==22.4.0
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
w3lib==1.22.0
zope.interface==5.4.0
